Allow custom IDs in scheduled workflow
