+++ lib/zenaton/services/graph_ql/dispatch_workflow_mutation.rb
@@ -1,15 +1,14 @@
 # frozen_string_literal: true
 
 require 'zenaton/services/graph_ql/base_operation'
-require 'zenaton/workflows/version'
-require 'zenaton/exceptions'
+require 'zenaton/concerns/workflow'
 
 module Zenaton
   module Services
     module GraphQL
       # Mutation parameters for executing a workflow
       class DispatchWorkflowMutation < BaseOperation
-        MAX_ID_SIZE = 256 # Maximum length for custom ids
+        include Concerns::Workflow
 
         # @raise [Zenaton::InvalidArgumentError] if custom id fails validation
         def initialize(workflow, app_env)
@@ -53,38 +52,6 @@ module Zenaton
           }
         end
         # rubocop:enable Metrics/MethodLength
-
-        private
-
-        def workflow_name
-          if @workflow.is_a? Workflows::Version
-            @workflow.current_implementation.class.name
-          else
-            @workflow.class.name
-          end
-        end
-
-        def validate_custom_id
-          return unless @workflow.try(:id).present?
-
-          validate_custom_id_type
-          validate_custom_id_value
-        end
-
-        def validate_custom_id_type
-          valid_types = [String, Integer]
-          return if valid_types.any? { |type| @workflow.id.is_a?(type) }
-
-          raise InvalidArgumentError,
-                'Provided ID must be a string or an integer' \
-        end
-
-        def validate_custom_id_value
-          return if @workflow.id.to_s.length <= MAX_ID_SIZE
-
-          raise InvalidArgumentError,
-                "Provided Id must not exceed #{MAX_ID_SIZE} bytes"
-        end
       end
     end
   end
