+++ spec/zenaton/services/graph_ql/create_workflow_schedule_mutation_spec.rb
@@ -2,12 +2,15 @@
 
 require 'zenaton/services/graph_ql/create_workflow_schedule_mutation'
 require 'fixtures/version'
+require 'shared_examples/mutation_with_custom_id'
 
 RSpec.describe Zenaton::Services::GraphQL::CreateWorkflowScheduleMutation do
   subject(:mutation) { described_class.new(workflow, cron, app_env) }
 
   let(:app_env) { 'dev' }
   let(:cron) { '* * * * *' }
+
+  it_behaves_like 'Mutation with CustomId'
 
   describe 'variables' do
     subject { mutation.variables }
@@ -20,6 +23,7 @@ RSpec.describe Zenaton::Services::GraphQL::CreateWorkflowScheduleMutation do
             'intentId' => String,
             'environmentName' => 'dev',
             'cron' => '* * * * *',
+            'customId' => nil,
             'workflowName' => 'FakeWorkflow2',
             'canonicalName' => 'FakeWorkflow2',
             'programmingLanguage' => 'RUBY',
@@ -42,12 +46,36 @@ RSpec.describe Zenaton::Services::GraphQL::CreateWorkflowScheduleMutation do
             'intentId' => String,
             'environmentName' => 'dev',
             'cron' => '* * * * *',
+            'customId' => 'my-custom-id',
             'workflowName' => 'FakeWorkflowWithID',
             'canonicalName' => 'FakeWorkflowWithID',
             'programmingLanguage' => 'RUBY',
             'properties' => {
               o: '@zenaton#0',
               s: [{ a: { '@custom_id': 'my-custom-id' } }]
+            }.to_json
+          }
+        }
+      end
+
+      it { is_expected.to match(expected_variables) }
+    end
+
+    context 'with an integer custom id' do
+      let(:workflow) { FakeWorkflowWithID.new(123) }
+      let(:expected_variables) do
+        {
+          'input' => {
+            'intentId' => String,
+            'environmentName' => 'dev',
+            'cron' => '* * * * *',
+            'customId' => '123',
+            'workflowName' => 'FakeWorkflowWithID',
+            'canonicalName' => 'FakeWorkflowWithID',
+            'programmingLanguage' => 'RUBY',
+            'properties' => {
+              o: '@zenaton#0',
+              s: [{ a: { '@custom_id': 123 } }]
             }.to_json
           }
         }
@@ -64,6 +92,7 @@ RSpec.describe Zenaton::Services::GraphQL::CreateWorkflowScheduleMutation do
             'intentId' => String,
             'environmentName' => 'dev',
             'cron' => '* * * * *',
+            'customId' => nil,
             'workflowName' => 'FakeWorkflow2',
             'canonicalName' => 'FakeVersion',
             'programmingLanguage' => 'RUBY',
